requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
openai==1.44.1
pydantic==2.9.1
pydantic-core==2.23.3
python-dotenv==1.0.1
regex==2024.7.24
requests==2.32.3
sniffio==1.3.1
tiktoken==0.7.0
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.2
